refactor(cli): split library into modules
