+++ crates/cli/src/lib.rs
@@ -1,942 +1,49 @@
 // crates/cli/src/lib.rs
-use std::collections::{HashMap, HashSet};
+use std::collections::HashSet;
 use std::env;
-use std::ffi::OsString;
 use std::fs;
 use std::io::{self, Read, Write};
-use std::net::{IpAddr, TcpStream};
+use std::net::TcpStream;
 #[cfg(unix)]
 use std::os::unix::fs::PermissionsExt;
-use std::sync::Arc;
 
-use daemon::{parse_config_file, parse_module, Module};
 use std::path::{Path, PathBuf};
-use std::time::{Duration, Instant};
+use std::time::Duration;
 
 use clap::parser::ValueSource;
-use clap::{ArgAction, ArgMatches, Args, CommandFactory, FromArgMatches, Parser};
+use clap::{ArgMatches, FromArgMatches};
 
+pub mod daemon;
 mod formatter;
-use compress::{available_codecs, Codec};
-use encoding_rs::Encoding;
-pub use engine::EngineError;
-use engine::{pipe_sessions, sync, DeleteMode, IdMapper, Result, Stats, StrongHash, SyncOptions};
-use filters::{default_cvs_rules, parse_with_options, Matcher, Rule};
-pub use formatter::render_help;
-use logging::{
-    parse_escapes, progress_formatter, DebugFlag, InfoFlag, LogFormat, SubscriberConfig,
+pub mod options;
+mod utils;
+
+use crate::daemon::run_daemon;
+pub use daemon::spawn_daemon_session;
+pub use options::cli_command;
+use options::{ClientOpts, ProbeOpts};
+use utils::{
+    init_logging, parse_filters, parse_name_map, parse_remote_specs, parse_rsync_path, RemoteSpec,
 };
-use meta::{parse_chmod, parse_chown, parse_id_map, IdKind};
-use protocol::CharsetConv;
+pub use utils::{parse_iconv, parse_rsh, print_version_if_requested};
+
+use compress::{available_codecs, Codec};
+pub use engine::EngineError;
+use engine::{pipe_sessions, sync, DeleteMode, Result, Stats, StrongHash, SyncOptions};
+use filters::{default_cvs_rules, Matcher, Rule};
+pub use formatter::render_help;
+use logging::{parse_escapes, progress_formatter, InfoFlag};
+use meta::{parse_chmod, parse_chown, IdKind};
 #[cfg(feature = "acl")]
 use protocol::CAP_ACLS;
 #[cfg(feature = "xattr")]
 use protocol::CAP_XATTRS;
 use protocol::{negotiate_version, ExitCode, CAP_CODECS, LATEST_VERSION, SUPPORTED_PROTOCOLS, V30};
-use shell_words::split as shell_split;
-use transport::{
-    parse_sockopts, AddressFamily, RateLimitedTransport, SockOpt, SshStdioTransport, TcpTransport,
-    Transport,
-};
+use transport::{parse_sockopts, AddressFamily, RateLimitedTransport, SshStdioTransport};
 #[cfg(unix)]
 use users::get_user_by_uid;
 
 pub mod version;
-
-pub fn print_version_if_requested<I>(args: I) -> bool
-where
-    I: IntoIterator<Item = OsString>,
-{
-    let mut show_version = false;
-    let mut quiet = false;
-    for arg in args {
-        if arg == "--version" || arg == "-V" {
-            show_version = true;
-        } else if arg == "--quiet" || arg == "-q" {
-            quiet = true;
-        }
-    }
-    if show_version {
-        if !quiet {
-            println!("{}", version::render_version_lines().join("\n"));
-        }
-        true
-    } else {
-        false
-    }
-}
-
-fn parse_filters(s: &str, from0: bool) -> std::result::Result<Vec<Rule>, filters::ParseError> {
-    let mut v = HashSet::new();
-    parse_with_options(s, from0, &mut v, 0, None)
-}
-
-fn parse_duration(s: &str) -> std::result::Result<Duration, std::num::ParseIntError> {
-    Ok(Duration::from_secs(s.parse()?))
-}
-
-fn parse_nonzero_duration(s: &str) -> std::result::Result<Duration, String> {
-    let d = parse_duration(s).map_err(|e| e.to_string())?;
-    if d.as_secs() == 0 {
-        Err("value must be greater than 0".into())
-    } else {
-        Ok(d)
-    }
-}
-
-const SIZE_SUFFIXES: &[(char, u32)] = &[('k', 10), ('m', 20), ('g', 30)];
-
-fn parse_suffixed<T>(s: &str, shifts: &[(char, u32)]) -> std::result::Result<T, String>
-where
-    T: TryFrom<u64>,
-{
-    let s = s.trim();
-    if s == "0" {
-        return T::try_from(0).map_err(|_| "size overflow".to_string());
-    }
-    if let Some(last) = s.chars().last() {
-        if last.is_ascii_alphabetic() {
-            let num = s[..s.len() - 1].parse::<u64>().map_err(|e| e.to_string())?;
-            let shift = shifts
-                .iter()
-                .find(|(c, _)| last.eq_ignore_ascii_case(c))
-                .map(|(_, s)| *s)
-                .ok_or_else(|| format!("invalid size suffix: {last}"))?;
-            let mult = 1u64 << shift;
-            let val = num
-                .checked_mul(mult)
-                .ok_or_else(|| "size overflow".to_string())?;
-            return T::try_from(val).map_err(|_| "size overflow".to_string());
-        }
-    }
-    let val = s.parse::<u64>().map_err(|e| e.to_string())?;
-    T::try_from(val).map_err(|_| "size overflow".to_string())
-}
-
-fn parse_size<T>(s: &str) -> std::result::Result<T, String>
-where
-    T: TryFrom<u64>,
-{
-    parse_suffixed(s, SIZE_SUFFIXES)
-}
-
-fn parse_dparam(s: &str) -> std::result::Result<(String, String), String> {
-    let mut parts = s.splitn(2, '=');
-    let name = parts
-        .next()
-        .ok_or_else(|| "invalid dparam".to_string())?
-        .to_string();
-    let value = parts
-        .next()
-        .ok_or_else(|| "invalid dparam".to_string())?
-        .to_string();
-    Ok((name, value))
-}
-
-fn parse_bool(s: &str) -> std::result::Result<bool, String> {
-    if ["1", "true", "yes"]
-        .iter()
-        .any(|v| s.eq_ignore_ascii_case(v))
-    {
-        Ok(true)
-    } else if ["0", "false", "no"]
-        .iter()
-        .any(|v| s.eq_ignore_ascii_case(v))
-    {
-        Ok(false)
-    } else {
-        Err("invalid boolean".to_string())
-    }
-}
-
-pub fn parse_logging_flags(matches: &ArgMatches) -> (Vec<InfoFlag>, Vec<DebugFlag>) {
-    let mut info: Vec<InfoFlag> = matches
-        .get_many::<InfoFlag>("info")
-        .map(|v| v.copied().collect())
-        .unwrap_or_default();
-    if matches.contains_id("out_format") && !info.contains(&InfoFlag::Name) {
-        info.push(InfoFlag::Name);
-    }
-    let debug = matches
-        .get_many::<DebugFlag>("debug")
-        .map(|v| v.copied().collect())
-        .unwrap_or_default();
-    (info, debug)
-}
-
-fn init_logging(matches: &ArgMatches) {
-    let verbose = matches.get_count("verbose");
-    let quiet = matches.get_flag("quiet");
-    let log_format = *matches
-        .get_one::<LogFormat>("log_format")
-        .unwrap_or(&LogFormat::Text);
-    let log_file = matches.get_one::<PathBuf>("client-log-file").cloned();
-    let log_file_fmt = matches.get_one::<String>("client-log-file-format").cloned();
-    let syslog = matches.get_flag("syslog");
-    let journald = matches.get_flag("journald");
-    let (mut info, mut debug) = parse_logging_flags(matches);
-    if quiet {
-        info.clear();
-        debug.clear();
-    }
-    let cfg = SubscriberConfig::builder()
-        .format(log_format)
-        .verbose(verbose)
-        .info(info)
-        .debug(debug)
-        .quiet(quiet)
-        .log_file(log_file.map(|p| (p, log_file_fmt)))
-        .syslog(syslog)
-        .journald(journald)
-        .colored(true)
-        .timestamps(false)
-        .build();
-    logging::init(cfg);
-}
-
-fn locale_charset() -> Option<String> {
-    for var in ["LC_ALL", "LC_CTYPE", "LANG"] {
-        if let Ok(val) = env::var(var) {
-            if let Some(enc) = val.split('.').nth(1) {
-                return Some(enc.to_string());
-            }
-        }
-    }
-    None
-}
-
-pub fn parse_iconv(spec: &str) -> std::result::Result<CharsetConv, String> {
-    let mut parts = spec.split(',');
-    let local_label = parts
-        .next()
-        .ok_or_else(|| "invalid iconv spec".to_string())?;
-    let remote_label = parts.next().unwrap_or("UTF-8");
-
-    let local_label = if local_label == "." {
-        locale_charset().ok_or_else(|| "failed to determine locale charset".to_string())?
-    } else {
-        local_label.to_string()
-    };
-    let remote_label = if remote_label == "." {
-        locale_charset().ok_or_else(|| "failed to determine locale charset".to_string())?
-    } else {
-        remote_label.to_string()
-    };
-
-    let local_enc = Encoding::for_label(local_label.as_bytes());
-    let remote_enc = Encoding::for_label(remote_label.as_bytes());
-
-    let local_enc = local_enc
-        .ok_or_else(|| format!("iconv_open(\"{local_label}\", \"{remote_label}\") failed"))?;
-    let remote_enc = remote_enc
-        .ok_or_else(|| format!("iconv_open(\"{local_label}\", \"{remote_label}\") failed"))?;
-
-    Ok(CharsetConv::new(remote_enc, local_enc))
-}
-
-#[derive(Parser, Debug)]
-struct ClientOpts {
-    #[arg(long)]
-    local: bool,
-    #[command(flatten)]
-    daemon: DaemonOpts,
-    #[arg(short = 'a', long, help_heading = "Selection")]
-    archive: bool,
-    #[arg(short, long, help_heading = "Selection")]
-    recursive: bool,
-    #[arg(short = 'd', long, help_heading = "Selection")]
-    dirs: bool,
-    #[arg(short = 'R', long, help_heading = "Selection")]
-    relative: bool,
-    #[arg(long = "no-implied-dirs", help_heading = "Selection")]
-    no_implied_dirs: bool,
-    #[arg(short = 'n', long, help_heading = "Selection")]
-    dry_run: bool,
-    #[arg(long = "list-only", help_heading = "Output")]
-    list_only: bool,
-    #[arg(short = 'S', long, help_heading = "Selection")]
-    sparse: bool,
-    #[arg(short = 'u', long, help_heading = "Misc")]
-    update: bool,
-    #[arg(long, help_heading = "Misc")]
-    existing: bool,
-    #[arg(long, help_heading = "Misc")]
-    ignore_existing: bool,
-    #[arg(short = 'x', long = "one-file-system", help_heading = "Selection")]
-    one_file_system: bool,
-    #[arg(short = 'm', long = "prune-empty-dirs", help_heading = "Misc")]
-    prune_empty_dirs: bool,
-    #[arg(long = "size-only", help_heading = "Misc")]
-    size_only: bool,
-    #[arg(short = 'I', long = "ignore-times", help_heading = "Misc")]
-    ignore_times: bool,
-    #[arg(short, long, action = ArgAction::Count, help_heading = "Output")]
-    verbose: u8,
-    #[arg(long = "log-format", help_heading = "Output", value_enum)]
-    log_format: Option<LogFormat>,
-    #[arg(
-        long = "log-file",
-        value_name = "FILE",
-        help_heading = "Output",
-        id = "client-log-file"
-    )]
-    log_file: Option<PathBuf>,
-    #[arg(
-        long = "log-file-format",
-        value_name = "FMT",
-        help_heading = "Output",
-        id = "client-log-file-format"
-    )]
-    log_file_format: Option<String>,
-    #[arg(long, help_heading = "Output", env = "OC_RSYNC_SYSLOG")]
-    syslog: bool,
-    #[arg(long, help_heading = "Output", env = "OC_RSYNC_JOURNALD")]
-    journald: bool,
-    #[arg(long = "out-format", value_name = "FORMAT", help_heading = "Output")]
-    out_format: Option<String>,
-    #[arg(
-        long,
-        value_name = "FLAGS",
-        value_delimiter = ',',
-        value_enum,
-        help_heading = "Output"
-    )]
-    info: Vec<InfoFlag>,
-    #[arg(
-        long,
-        value_name = "FLAGS",
-        value_delimiter = ',',
-        value_enum,
-        help_heading = "Output"
-    )]
-    debug: Vec<DebugFlag>,
-    #[arg(long = "human-readable", help_heading = "Output")]
-    human_readable: bool,
-    #[arg(short, long, help_heading = "Output")]
-    quiet: bool,
-    #[arg(long, help_heading = "Output")]
-    no_motd: bool,
-    #[arg(short = '8', long = "8-bit-output", help_heading = "Output")]
-    eight_bit_output: bool,
-    #[arg(
-        short = 'i',
-        long = "itemize-changes",
-        help_heading = "Output",
-        help = "output a change-summary for all updates"
-    )]
-    itemize_changes: bool,
-    #[arg(
-        long,
-        help_heading = "Delete",
-        overrides_with_all = [
-            "delete_before",
-            "delete_during",
-            "delete_after",
-            "delete_delay"
-        ]
-    )]
-    delete: bool,
-    #[arg(
-        long = "delete-before",
-        help_heading = "Delete",
-        overrides_with_all = ["delete", "delete_during", "delete_after", "delete_delay"]
-    )]
-    delete_before: bool,
-    #[arg(
-        long = "delete-during",
-        help_heading = "Delete",
-        visible_alias = "del",
-        overrides_with_all = ["delete", "delete_before", "delete_after", "delete_delay"]
-    )]
-    delete_during: bool,
-    #[arg(
-        long = "delete-after",
-        help_heading = "Delete",
-        overrides_with_all = ["delete", "delete_before", "delete_during", "delete_delay"]
-    )]
-    delete_after: bool,
-    #[arg(
-        long = "delete-delay",
-        help_heading = "Delete",
-        overrides_with_all = ["delete", "delete_before", "delete_during", "delete_after"]
-    )]
-    delete_delay: bool,
-    #[arg(long = "delete-excluded", help_heading = "Delete")]
-    delete_excluded: bool,
-    #[arg(long = "delete-missing-args", help_heading = "Delete")]
-    delete_missing_args: bool,
-    #[arg(long = "ignore-missing-args", help_heading = "Delete")]
-    ignore_missing_args: bool,
-    #[arg(
-        long = "remove-source-files",
-        help_heading = "Delete",
-        visible_alias = "remove-sent-files"
-    )]
-    remove_source_files: bool,
-    #[arg(long = "ignore-errors", help_heading = "Delete")]
-    ignore_errors: bool,
-    #[arg(
-        long,
-        help_heading = "Delete",
-        help = "force deletion of dirs even if not empty",
-        action = ArgAction::SetTrue
-    )]
-    force: bool,
-    #[arg(long = "max-delete", value_name = "NUM", help_heading = "Delete")]
-    max_delete: Option<usize>,
-    #[arg(
-        long = "max-alloc",
-        value_name = "SIZE",
-        value_parser = parse_size::<usize>,
-        help_heading = "Misc"
-    )]
-    max_alloc: Option<usize>,
-    #[arg(
-        long = "max-size",
-        value_name = "SIZE",
-        value_parser = parse_size::<u64>,
-        help_heading = "Misc"
-    )]
-    max_size: Option<u64>,
-    #[arg(
-        long = "min-size",
-        value_name = "SIZE",
-        value_parser = parse_size::<u64>,
-        help_heading = "Misc"
-    )]
-    min_size: Option<u64>,
-    #[arg(
-        long,
-        help_heading = "Misc",
-        help = "allocate dest files before writing them"
-    )]
-    preallocate: bool,
-    #[arg(short = 'b', long, help_heading = "Backup")]
-    backup: bool,
-    #[arg(long = "backup-dir", value_name = "DIR", help_heading = "Backup")]
-    backup_dir: Option<PathBuf>,
-    #[arg(
-        long = "suffix",
-        value_name = "SUFFIX",
-        default_value = "~",
-        help_heading = "Backup"
-    )]
-    suffix: String,
-    #[arg(short = 'c', long, help_heading = "Attributes")]
-    checksum: bool,
-    #[arg(
-        long = "checksum-choice",
-        value_name = "STR",
-        help_heading = "Attributes",
-        visible_alias = "cc"
-    )]
-    checksum_choice: Option<String>,
-    #[arg(
-        long = "checksum-seed",
-        value_name = "NUM",
-        value_parser = clap::value_parser!(u32),
-        help_heading = "Attributes",
-        help = "set block/file checksum seed (advanced)"
-    )]
-    checksum_seed: Option<u32>,
-    #[arg(
-        short = 'p',
-        long,
-        help_heading = "Attributes",
-        overrides_with = "no_perms"
-    )]
-    perms: bool,
-    #[arg(
-        long = "no-perms",
-        help_heading = "Attributes",
-        overrides_with = "perms"
-    )]
-    no_perms: bool,
-    #[arg(short = 'E', long, help_heading = "Attributes")]
-    executability: bool,
-    #[arg(long = "chmod", value_name = "CHMOD", help_heading = "Attributes")]
-    chmod: Vec<String>,
-    #[arg(long = "chown", value_name = "USER:GROUP", help_heading = "Attributes")]
-    chown: Option<String>,
-    #[arg(
-        long = "copy-as",
-        value_name = "USER[:GROUP]",
-        help_heading = "Attributes"
-    )]
-    copy_as: Option<String>,
-    #[arg(
-        long = "usermap",
-        value_name = "FROM:TO",
-        value_delimiter = ',',
-        help_heading = "Attributes"
-    )]
-    usermap: Vec<String>,
-    #[arg(
-        long = "groupmap",
-        value_name = "FROM:TO",
-        value_delimiter = ',',
-        help_heading = "Attributes"
-    )]
-    groupmap: Vec<String>,
-    #[arg(
-        short = 't',
-        long,
-        help_heading = "Attributes",
-        overrides_with = "no_times"
-    )]
-    times: bool,
-    #[arg(
-        long = "no-times",
-        help_heading = "Attributes",
-        overrides_with = "times"
-    )]
-    no_times: bool,
-    #[arg(short = 'U', long, help_heading = "Attributes")]
-    atimes: bool,
-    #[arg(short = 'N', long, help_heading = "Attributes")]
-    crtimes: bool,
-    #[arg(short = 'O', long, help_heading = "Attributes")]
-    omit_dir_times: bool,
-    #[arg(short = 'J', long, help_heading = "Attributes")]
-    omit_link_times: bool,
-    #[arg(
-        short = 'o',
-        long,
-        help_heading = "Attributes",
-        overrides_with = "no_owner"
-    )]
-    owner: bool,
-    #[arg(
-        long = "no-owner",
-        alias = "no-o",
-        help_heading = "Attributes",
-        overrides_with = "owner",
-        alias = "no-o"
-    )]
-    no_owner: bool,
-    #[arg(
-        short = 'g',
-        long,
-        help_heading = "Attributes",
-        overrides_with = "no_group"
-    )]
-    group: bool,
-    #[arg(
-        long = "no-group",
-        help_heading = "Attributes",
-        overrides_with = "group",
-        alias = "no-g"
-    )]
-    no_group: bool,
-    #[arg(
-        short = 'l',
-        long,
-        help_heading = "Attributes",
-        overrides_with = "no_links"
-    )]
-    links: bool,
-    #[arg(
-        long = "no-links",
-        help_heading = "Attributes",
-        overrides_with = "links"
-    )]
-    no_links: bool,
-    #[arg(short = 'L', long, help_heading = "Attributes")]
-    copy_links: bool,
-    #[arg(short = 'k', long, help_heading = "Attributes")]
-    copy_dirlinks: bool,
-    #[arg(short = 'K', long, help_heading = "Attributes")]
-    keep_dirlinks: bool,
-    #[arg(long, help_heading = "Attributes")]
-    copy_unsafe_links: bool,
-    #[arg(long, help_heading = "Attributes")]
-    safe_links: bool,
-    #[arg(
-        long,
-        help_heading = "Attributes",
-        help = "munge symlinks to make them safe & unusable"
-    )]
-    munge_links: bool,
-    #[arg(short = 'H', long = "hard-links", help_heading = "Attributes")]
-    hard_links: bool,
-    #[arg(long, help_heading = "Attributes", overrides_with = "no_devices")]
-    devices: bool,
-    #[arg(
-        long = "no-devices",
-        help_heading = "Attributes",
-        overrides_with = "devices"
-    )]
-    no_devices: bool,
-    #[arg(long, help_heading = "Attributes", overrides_with = "no_specials")]
-    specials: bool,
-    #[arg(
-        long = "no-specials",
-        help_heading = "Attributes",
-        overrides_with = "specials"
-    )]
-    no_specials: bool,
-    #[arg(short = 'D', help_heading = "Attributes")]
-    devices_specials: bool,
-    #[cfg(feature = "xattr")]
-    #[arg(long, help_heading = "Attributes")]
-    xattrs: bool,
-    #[cfg(feature = "acl")]
-    #[arg(
-        short = 'A',
-        long,
-        help_heading = "Attributes",
-        overrides_with = "no_acls"
-    )]
-    acls: bool,
-    #[cfg(feature = "acl")]
-    #[arg(long = "no-acls", help_heading = "Attributes", overrides_with = "acls")]
-    no_acls: bool,
-    #[arg(long = "fake-super", help_heading = "Attributes")]
-    fake_super: bool,
-    #[arg(long = "super", help_heading = "Attributes")]
-    super_user: bool,
-    #[arg(short = 'z', long, help_heading = "Compression")]
-    compress: bool,
-    #[arg(
-        long = "compress-choice",
-        value_name = "STR",
-        help_heading = "Compression",
-        visible_alias = "zc"
-    )]
-    compress_choice: Option<String>,
-    #[arg(
-        long = "compress-level",
-        value_name = "NUM",
-        help_heading = "Compression",
-        visible_alias = "zl"
-    )]
-    compress_level: Option<i32>,
-    #[arg(
-        long = "skip-compress",
-        value_name = "LIST",
-        help_heading = "Compression",
-        value_delimiter = ','
-    )]
-    skip_compress: Vec<String>,
-
-    #[arg(long, help_heading = "Misc")]
-    partial: bool,
-    #[arg(long = "partial-dir", value_name = "DIR", help_heading = "Misc")]
-    partial_dir: Option<PathBuf>,
-    #[arg(
-        short = 'T',
-        long = "temp-dir",
-        value_name = "DIR",
-        help_heading = "Misc"
-    )]
-    temp_dir: Option<PathBuf>,
-    #[arg(long, help_heading = "Misc")]
-    progress: bool,
-    #[arg(long, help_heading = "Misc")]
-    blocking_io: bool,
-    #[arg(long, help_heading = "Misc")]
-    fsync: bool,
-    #[arg(short = 'y', long = "fuzzy", help_heading = "Misc")]
-    fuzzy: bool,
-    #[arg(short = 'P', help_heading = "Misc")]
-    partial_progress: bool,
-    #[arg(long, help_heading = "Misc")]
-    append: bool,
-    #[arg(long = "append-verify", help_heading = "Misc")]
-    append_verify: bool,
-    #[arg(long, help_heading = "Misc")]
-    inplace: bool,
-    #[arg(long = "delay-updates", help_heading = "Misc")]
-    delay_updates: bool,
-    #[arg(long = "bwlimit", value_name = "RATE", help_heading = "Misc")]
-    bwlimit: Option<u64>,
-    #[arg(long = "timeout", value_name = "SECONDS", value_parser = parse_duration, help_heading = "Misc")]
-    timeout: Option<Duration>,
-    #[arg(
-        long = "connect-timeout",
-        alias = "contimeout",
-        value_name = "SECONDS",
-        value_parser = parse_nonzero_duration,
-        help_heading = "Misc"
-    )]
-    connect_timeout: Option<Duration>,
-    #[arg(long = "modify-window", value_name = "SECONDS", value_parser = parse_duration, help_heading = "Misc")]
-    modify_window: Option<Duration>,
-    #[arg(
-        long = "protocol",
-        value_name = "VER",
-        value_parser = clap::value_parser!(u32),
-        help_heading = "Misc",
-        help = "force an older protocol version"
-    )]
-    protocol: Option<u32>,
-    #[arg(long, value_name = "PORT", help_heading = "Misc")]
-    port: Option<u16>,
-    #[arg(
-        short = '4',
-        long = "ipv4",
-        help_heading = "Misc",
-        conflicts_with = "ipv6"
-    )]
-    ipv4: bool,
-    #[arg(
-        short = '6',
-        long = "ipv6",
-        help_heading = "Misc",
-        conflicts_with = "ipv4"
-    )]
-    ipv6: bool,
-    #[arg(
-        short = 'B',
-        long = "block-size",
-        value_name = "SIZE",
-        help_heading = "Misc"
-    )]
-    block_size: Option<usize>,
-    #[arg(
-        short = 'W',
-        long,
-        help_heading = "Misc",
-        overrides_with = "no_whole_file"
-    )]
-    whole_file: bool,
-    #[arg(
-        long = "no-whole-file",
-        help_heading = "Misc",
-        overrides_with = "whole_file"
-    )]
-    no_whole_file: bool,
-    #[arg(long = "link-dest", value_name = "DIR", help_heading = "Misc")]
-    link_dest: Option<PathBuf>,
-    #[arg(long = "copy-dest", value_name = "DIR", help_heading = "Misc")]
-    copy_dest: Option<PathBuf>,
-    #[arg(long = "compare-dest", value_name = "DIR", help_heading = "Misc")]
-    compare_dest: Option<PathBuf>,
-    #[arg(
-        long = "mkpath",
-        help_heading = "Misc",
-        help = "create destination's missing path components"
-    )]
-    mkpath: bool,
-    #[arg(long, help_heading = "Attributes")]
-    numeric_ids: bool,
-    #[arg(long, help_heading = "Output")]
-    stats: bool,
-    #[arg(long, value_name = "FILE")]
-    config: Option<PathBuf>,
-    #[arg(long, value_name = "FILE", env = "RSYNC_KNOWN_HOSTS")]
-    known_hosts: Option<PathBuf>,
-    #[arg(long, env = "RSYNC_NO_HOST_KEY_CHECKING")]
-    no_host_key_checking: bool,
-    #[arg(long = "password-file", value_name = "FILE")]
-    password_file: Option<PathBuf>,
-    #[arg(long = "early-input", value_name = "FILE")]
-    early_input: Option<PathBuf>,
-    #[arg(short = 'e', long, value_name = "COMMAND")]
-    rsh: Option<String>,
-    #[arg(
-        short = 'M',
-        long = "remote-option",
-        value_name = "OPT",
-        allow_hyphen_values = true,
-        help = "send OPTION to the remote side only"
-    )]
-    remote_option: Vec<String>,
-    #[arg(
-        short = 's',
-        long = "secluded-args",
-        help_heading = "Misc",
-        help = "use the protocol to safely send the args"
-    )]
-    secluded_args: bool,
-    #[arg(long = "trust-sender", help_heading = "Misc")]
-    trust_sender: bool,
-    #[arg(
-        long = "sockopts",
-        value_name = "OPTIONS",
-        value_delimiter = ',',
-        allow_hyphen_values = true,
-        help_heading = "Misc"
-    )]
-    sockopts: Vec<String>,
-    #[arg(
-        long = "iconv",
-        value_name = "CONVERT_SPEC",
-        help_heading = "Misc",
-        help = "request charset conversion of filenames"
-    )]
-    iconv: Option<String>,
-    #[arg(
-        long = "write-batch",
-        value_name = "FILE",
-        help_heading = "Misc",
-        conflicts_with = "read_batch"
-    )]
-    write_batch: Option<PathBuf>,
-    #[arg(
-        long = "read-batch",
-        value_name = "FILE",
-        help_heading = "Misc",
-        help = "read a batched update from FILE",
-        conflicts_with = "write_batch"
-    )]
-    read_batch: Option<PathBuf>,
-    #[arg(long = "copy-devices", help_heading = "Misc")]
-    copy_devices: bool,
-    #[arg(
-        long = "write-devices",
-        help = "write to devices as files (implies --inplace)",
-        help_heading = "Misc"
-    )]
-    write_devices: bool,
-    #[arg(long, hide = true)]
-    server: bool,
-    #[arg(long, hide = true)]
-    sender: bool,
-    #[arg(long = "rsync-path", value_name = "PATH", alias = "rsync_path")]
-    rsync_path: Option<String>,
-    #[arg(value_name = "SRC", required_unless_present_any = ["daemon", "server", "probe"])]
-    src: Option<String>,
-    #[arg(value_name = "DST", required_unless_present_any = ["daemon", "server", "probe"])]
-    dst: Option<String>,
-    #[arg(short = 'f', long, value_name = "RULE", help_heading = "Selection")]
-    filter: Vec<String>,
-    #[arg(long, value_name = "FILE", help_heading = "Selection")]
-    filter_file: Vec<PathBuf>,
-    #[arg(short = 'F', action = ArgAction::Count, help_heading = "Selection")]
-    filter_shorthand: u8,
-    #[arg(
-        short = 'C',
-        long = "cvs-exclude",
-        help_heading = "Selection",
-        help = "auto-ignore files in the same way CVS does"
-    )]
-    cvs_exclude: bool,
-    #[arg(long, value_name = "PATTERN")]
-    include: Vec<String>,
-    #[arg(long, value_name = "PATTERN")]
-    exclude: Vec<String>,
-    #[arg(long, value_name = "FILE")]
-    include_from: Vec<PathBuf>,
-    #[arg(long, value_name = "FILE")]
-    exclude_from: Vec<PathBuf>,
-    #[arg(long, value_name = "FILE")]
-    files_from: Vec<PathBuf>,
-    #[arg(long, short = '0')]
-    from0: bool,
-}
-
-#[doc(hidden)]
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct RshCommand {
-    pub env: Vec<(String, String)>,
-    pub cmd: Vec<String>,
-}
-
-fn parse_env_command(parts: Vec<String>) -> Result<RshCommand> {
-    let mut env = Vec::new();
-    let mut iter = parts.into_iter();
-    let mut cmd = Vec::new();
-
-    while let Some(tok) = iter.next() {
-        if let Some((k, v)) = tok.split_once('=') {
-            let valid = !k.is_empty()
-                && (k.as_bytes()[0].is_ascii_alphabetic() || k.as_bytes()[0] == b'_')
-                && k.as_bytes()[1..]
-                    .iter()
-                    .all(|b| b.is_ascii_alphanumeric() || *b == b'_');
-            if valid {
-                env.push((k.to_string(), v.to_string()));
-                continue;
-            } else {
-                return Err(EngineError::Other("invalid environment assignment".into()));
-            }
-        }
-        cmd.push(tok);
-        cmd.extend(iter);
-        return Ok(RshCommand { env, cmd });
-    }
-    Ok(RshCommand { env, cmd })
-}
-
-pub fn parse_rsh(raw: Option<String>) -> Result<RshCommand> {
-    match raw {
-        Some(s) => {
-            let parts = shell_split(&s).map_err(|e| EngineError::Other(e.to_string()))?;
-            let mut cmd = parse_env_command(parts)?;
-            if cmd.cmd.is_empty() {
-                cmd.cmd.push("ssh".to_string());
-            }
-            Ok(cmd)
-        }
-        None => Ok(RshCommand {
-            env: Vec::new(),
-            cmd: vec!["ssh".to_string()],
-        }),
-    }
-}
-
-pub fn parse_rsync_path(raw: Option<String>) -> Result<Option<RshCommand>> {
-    match raw {
-        Some(s) => {
-            let parts = shell_split(&s).map_err(|e| EngineError::Other(e.to_string()))?;
-            let cmd = parse_env_command(parts)?;
-            if cmd.env.is_empty() && cmd.cmd.is_empty() {
-                Ok(None)
-            } else {
-                Ok(Some(cmd))
-            }
-        }
-        None => Ok(None),
-    }
-}
-
-fn parse_name_map(specs: &[String], kind: IdKind) -> Result<Option<IdMapper>> {
-    if specs.is_empty() {
-        Ok(None)
-    } else {
-        let spec = specs.join(",");
-        let mapper = parse_id_map(&spec, kind).map_err(EngineError::Other)?;
-        Ok(Some(IdMapper(mapper)))
-    }
-}
-
-#[derive(Args, Debug)]
-struct DaemonOpts {
-    #[arg(long)]
-    daemon: bool,
-    #[arg(long = "no-detach")]
-    no_detach: bool,
-    #[arg(long, value_parser = parse_module, value_name = "NAME=PATH")]
-    module: Vec<Module>,
-    #[arg(long)]
-    address: Option<IpAddr>,
-    #[arg(long = "secrets-file", value_name = "FILE")]
-    secrets_file: Option<PathBuf>,
-    #[arg(long = "hosts-allow", value_delimiter = ',', value_name = "LIST")]
-    hosts_allow: Vec<String>,
-    #[arg(long = "hosts-deny", value_delimiter = ',', value_name = "LIST")]
-    hosts_deny: Vec<String>,
-    #[arg(long = "motd", value_name = "FILE")]
-    motd: Option<PathBuf>,
-    #[arg(long = "pid-file", value_name = "FILE")]
-    pid_file: Option<PathBuf>,
-    #[arg(long = "lock-file", value_name = "FILE")]
-    lock_file: Option<PathBuf>,
-    #[arg(long = "state-dir", value_name = "DIR")]
-    state_dir: Option<PathBuf>,
-    #[arg(long = "dparam", value_name = "NAME=VALUE", value_parser = parse_dparam)]
-    dparam: Vec<(String, String)>,
-}
-
-#[derive(Parser, Debug)]
-struct ProbeOpts {
-    #[arg(long)]
-    probe: bool,
-    addr: Option<String>,
-    #[arg(long, default_value_t = SUPPORTED_PROTOCOLS[0], value_name = "VER")]
-    peer_version: u32,
-}
 
 pub fn run(matches: &clap::ArgMatches) -> Result<()> {
     init_logging(matches);
@@ -960,244 +67,6 @@ pub fn run(matches: &clap::ArgMatches) -> Result<()> {
         }
     }
     run_client(opts, matches)
-}
-
-pub fn cli_command() -> clap::Command {
-    let cmd = ClientOpts::command();
-    let cmd = ProbeOpts::augment_args(cmd);
-    formatter::apply(cmd)
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PathSpec {
-    path: PathBuf,
-    trailing_slash: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum RemoteSpec {
-    Local(PathSpec),
-    Remote {
-        host: String,
-        path: PathSpec,
-        module: Option<String>,
-    },
-}
-
-pub fn parse_remote_spec(input: &str) -> Result<RemoteSpec> {
-    let (trailing_slash, s) = if input != "/" && input.ends_with('/') {
-        (true, &input[..input.len() - 1])
-    } else {
-        (false, input)
-    };
-    if let Some(rest) = s.strip_prefix("rsync://") {
-        let mut parts = rest.splitn(2, '/');
-        let host = parts.next().unwrap_or("");
-        let mod_path = parts.next().unwrap_or("");
-        let mut mp = mod_path.splitn(2, '/');
-        let module = mp.next().unwrap_or("");
-        let path = mp.next().unwrap_or("");
-        return Ok(RemoteSpec::Remote {
-            host: host.to_string(),
-            path: PathSpec {
-                path: PathBuf::from(path),
-                trailing_slash,
-            },
-            module: Some(module.to_string()),
-        });
-    }
-    if let Some(rest) = s.strip_prefix('[') {
-        if let Some(end) = rest.find(']') {
-            let host = &rest[..end];
-            if let Some(path) = rest[end + 1..].strip_prefix(':') {
-                return Ok(RemoteSpec::Remote {
-                    host: host.to_string(),
-                    path: PathSpec {
-                        path: PathBuf::from(path),
-                        trailing_slash,
-                    },
-                    module: None,
-                });
-            }
-        }
-        return Ok(RemoteSpec::Local(PathSpec {
-            path: PathBuf::from(input),
-            trailing_slash,
-        }));
-    }
-    if let Some(idx) = s.find("::") {
-        let host = &s[..idx];
-        let mut rest = s[idx + 2..].splitn(2, '/');
-        let module = rest.next().unwrap_or("");
-        let path = rest.next().unwrap_or("");
-        return Ok(RemoteSpec::Remote {
-            host: host.to_string(),
-            path: PathSpec {
-                path: PathBuf::from(path),
-                trailing_slash,
-            },
-            module: Some(module.to_string()),
-        });
-    }
-    if let Some(idx) = s.find(':') {
-        if idx == 1 {
-            let bytes = s.as_bytes();
-            if bytes[0].is_ascii_alphabetic()
-                && (bytes.len() == 2
-                    || bytes
-                        .get(2)
-                        .map(|c| *c == b'/' || *c == b'\\')
-                        .unwrap_or(false))
-            {
-                return Ok(RemoteSpec::Local(PathSpec {
-                    path: PathBuf::from(s),
-                    trailing_slash,
-                }));
-            }
-        }
-        let (host, path) = s.split_at(idx);
-        return Ok(RemoteSpec::Remote {
-            host: host.to_string(),
-            path: PathSpec {
-                path: PathBuf::from(&path[1..]),
-                trailing_slash,
-            },
-            module: None,
-        });
-    }
-    Ok(RemoteSpec::Local(PathSpec {
-        path: PathBuf::from(s),
-        trailing_slash,
-    }))
-}
-
-pub fn parse_remote_specs(src: &str, dst: &str) -> Result<(RemoteSpec, RemoteSpec)> {
-    let src_spec = parse_remote_spec(src)?;
-    let dst_spec = parse_remote_spec(dst)?;
-    if let (
-        RemoteSpec::Remote {
-            host: sh, path: sp, ..
-        },
-        RemoteSpec::Remote {
-            host: dh, path: dp, ..
-        },
-    ) = (&src_spec, &dst_spec)
-    {
-        if sh.is_empty() || dh.is_empty() {
-            return Err(EngineError::Other("remote host missing".into()));
-        }
-        if sp.path.as_os_str().is_empty() || dp.path.as_os_str().is_empty() {
-            return Err(EngineError::Other("remote path missing".into()));
-        }
-    }
-    Ok((src_spec, dst_spec))
-}
-
-#[allow(clippy::too_many_arguments)]
-pub fn spawn_daemon_session(
-    host: &str,
-    module: &str,
-    port: Option<u16>,
-    password_file: Option<&Path>,
-    no_motd: bool,
-    timeout: Option<Duration>,
-    connect_timeout: Option<Duration>,
-    family: Option<AddressFamily>,
-    sockopts: &[String],
-    opts: &SyncOptions,
-    version: u32,
-    early_input: Option<&Path>,
-    iconv: Option<&CharsetConv>,
-) -> Result<TcpTransport> {
-    let (host, port) = if let Some((h, p)) = host.rsplit_once(':') {
-        let p = p.parse().unwrap_or(873);
-        (h, p)
-    } else {
-        (host, port.unwrap_or(873))
-    };
-    let start = Instant::now();
-    let mut t =
-        TcpTransport::connect(host, port, connect_timeout, family).map_err(EngineError::from)?;
-    let parsed: Vec<SockOpt> = parse_sockopts(sockopts).map_err(EngineError::Other)?;
-    t.apply_sockopts(&parsed).map_err(EngineError::from)?;
-    let handshake_timeout = connect_timeout
-        .map(|dur| {
-            dur.checked_sub(start.elapsed())
-                .ok_or_else(|| io::Error::new(io::ErrorKind::TimedOut, "connection timed out"))
-        })
-        .transpose()
-        .map_err(EngineError::from)?;
-    t.set_read_timeout(handshake_timeout)
-        .map_err(EngineError::from)?;
-    t.set_write_timeout(handshake_timeout)
-        .map_err(EngineError::from)?;
-    if let Some(p) = early_input {
-        if let Ok(data) = fs::read(p) {
-            t.send(&data).map_err(EngineError::from)?;
-        }
-    }
-    t.send(&version.to_be_bytes()).map_err(EngineError::from)?;
-    let mut buf = [0u8; 4];
-    t.receive(&mut buf).map_err(EngineError::from)?;
-    let peer = u32::from_be_bytes(buf);
-    negotiate_version(version, peer).map_err(|e| EngineError::Other(e.to_string()))?;
-
-    let token = password_file
-        .and_then(|p| fs::read_to_string(p).ok())
-        .and_then(|s| s.lines().next().map(|l| l.to_string()));
-    t.authenticate(token.as_deref(), no_motd)
-        .map_err(EngineError::from)?;
-
-    let mut line = Vec::new();
-    let mut b = [0u8; 1];
-    loop {
-        let n = t.receive(&mut b).map_err(EngineError::from)?;
-        if n == 0 {
-            break;
-        }
-        line.push(b[0]);
-        if b[0] == b'\n' {
-            if line == b"@RSYNCD: OK\n" {
-                break;
-            }
-            if !no_motd {
-                let s = if let Some(cv) = iconv {
-                    cv.decode_remote(&line)
-                } else {
-                    String::from_utf8_lossy(&line).into_owned()
-                };
-                if let Some(msg) = s.strip_prefix("@RSYNCD: ") {
-                    print!("{msg}");
-                } else {
-                    print!("{s}");
-                }
-                let _ = io::stdout().flush();
-            }
-            line.clear();
-        }
-    }
-    t.set_read_timeout(timeout).map_err(EngineError::from)?;
-    t.set_write_timeout(timeout).map_err(EngineError::from)?;
-
-    if let Some(cv) = iconv {
-        let mut line = cv.encode_remote(module);
-        line.push(b'\n');
-        t.send(&line).map_err(EngineError::from)?;
-        for opt in &opts.remote_options {
-            let mut o = cv.encode_remote(opt);
-            o.push(b'\n');
-            t.send(&o).map_err(EngineError::from)?;
-        }
-    } else {
-        let line = format!("{module}\n");
-        t.send(line.as_bytes()).map_err(EngineError::from)?;
-        for opt in &opts.remote_options {
-            let o = format!("{opt}\n");
-            t.send(o.as_bytes()).map_err(EngineError::from)?;
-        }
-    }
-    t.send(b"\n").map_err(EngineError::from)?;
-    Ok(t)
 }
 
 fn run_client(mut opts: ClientOpts, matches: &ArgMatches) -> Result<()> {
@@ -2363,214 +1232,6 @@ fn build_matcher(opts: &ClientOpts, matches: &ArgMatches) -> Result<Matcher> {
     Ok(matcher)
 }
 
-fn run_daemon(opts: DaemonOpts, matches: &ArgMatches) -> Result<()> {
-    let mut modules: HashMap<String, Module> = HashMap::new();
-    let mut secrets = opts.secrets_file.clone();
-    let password = matches
-        .get_one::<PathBuf>("password_file")
-        .cloned()
-        .map(|pf| -> Result<String> {
-            #[cfg(unix)]
-            {
-                let mode = fs::metadata(&pf)?.permissions().mode();
-                if mode & 0o077 != 0 {
-                    return Err(EngineError::Other(
-                        "password file permissions are too open".into(),
-                    ));
-                }
-            }
-            let data = fs::read_to_string(&pf)?;
-            Ok(data.lines().next().unwrap_or_default().trim().to_string())
-        })
-        .transpose()?;
-    let mut hosts_allow = opts.hosts_allow.clone();
-    let mut hosts_deny = opts.hosts_deny.clone();
-    let mut log_file = matches.get_one::<PathBuf>("client-log-file").cloned();
-    let log_format = matches
-        .get_one::<String>("client-log-file-format")
-        .map(|s| parse_escapes(s));
-    let mut motd = opts.motd.clone();
-    let mut pid_file = opts.pid_file.clone();
-    let mut lock_file = opts.lock_file.clone();
-    let mut state_dir = opts.state_dir.clone();
-    let mut port = matches.get_one::<u16>("port").copied().unwrap_or(873);
-    let mut address = opts.address;
-    let timeout = matches.get_one::<Duration>("timeout").copied();
-    let bwlimit = matches.get_one::<u64>("bwlimit").copied();
-    let numeric_ids_flag = matches.get_flag("numeric_ids");
-    let mut list = true;
-    let mut refuse = Vec::new();
-    let mut max_conn = None;
-    let mut read_only = None;
-    if let Some(cfg_path) = matches.get_one::<PathBuf>("config") {
-        let cfg = parse_config_file(cfg_path).map_err(|e| EngineError::Other(e.to_string()))?;
-        for m in cfg.modules {
-            modules.insert(m.name.clone(), m);
-        }
-        if let Some(p) = cfg.port {
-            port = p;
-        }
-        if let Some(m) = cfg.motd_file {
-            motd = Some(m);
-        }
-        if let Some(l) = cfg.log_file {
-            log_file = Some(l);
-        }
-        if let Some(s) = cfg.secrets_file {
-            secrets = Some(s);
-        }
-        if let Some(p) = cfg.pid_file {
-            pid_file = Some(p);
-        }
-        if let Some(l) = cfg.lock_file {
-            lock_file = Some(l);
-        }
-        if let Some(a) = cfg.address {
-            address = Some(a);
-        }
-        if !cfg.hosts_allow.is_empty() {
-            hosts_allow = cfg.hosts_allow;
-        }
-        if !cfg.hosts_deny.is_empty() {
-            hosts_deny = cfg.hosts_deny;
-        }
-        if let Some(val) = cfg.numeric_ids {
-            for m in modules.values_mut() {
-                m.numeric_ids = val;
-            }
-        }
-        if let Some(val) = cfg.read_only {
-            read_only = Some(val);
-        }
-        if let Some(val) = cfg.list {
-            list = val;
-        }
-        if let Some(val) = cfg.max_connections {
-            max_conn = Some(val);
-        }
-        if !cfg.refuse_options.is_empty() {
-            refuse = cfg.refuse_options;
-        }
-    }
-
-    for m in opts.module {
-        modules.insert(m.name.clone(), m);
-    }
-
-    for (name, value) in opts.dparam {
-        match name.as_str() {
-            "motdfile" => motd = Some(PathBuf::from(value)),
-            "pidfile" => pid_file = Some(PathBuf::from(value)),
-            "logfile" => log_file = Some(PathBuf::from(value)),
-            "lockfile" => lock_file = Some(PathBuf::from(value)),
-            "statedir" => state_dir = Some(PathBuf::from(value)),
-            "secretsfile" => secrets = Some(PathBuf::from(value)),
-            "address" => {
-                address = Some(
-                    value
-                        .parse::<IpAddr>()
-                        .map_err(|e| EngineError::Other(e.to_string()))?,
-                )
-            }
-            "port" => {
-                port = value
-                    .parse::<u16>()
-                    .map_err(|e| EngineError::Other(e.to_string()))?
-            }
-            "numericids" => {
-                let val = parse_bool(&value).map_err(EngineError::Other)?;
-                for m in modules.values_mut() {
-                    m.numeric_ids = val;
-                }
-            }
-            "read only" | "read_only" => {
-                let val = parse_bool(&value).map_err(EngineError::Other)?;
-                for m in modules.values_mut() {
-                    m.read_only = val;
-                }
-            }
-            "list" => {
-                list = parse_bool(&value).map_err(EngineError::Other)?;
-            }
-            "max connections" | "maxconnections" => {
-                max_conn = Some(
-                    value
-                        .parse::<usize>()
-                        .map_err(|e| EngineError::Other(e.to_string()))?,
-                );
-            }
-            "hosts allow" | "hostsallow" => {
-                hosts_allow = value.split_whitespace().map(|s| s.to_string()).collect();
-            }
-            "hosts deny" | "hostsdeny" => {
-                hosts_deny = value.split_whitespace().map(|s| s.to_string()).collect();
-            }
-            "refuse options" | "refuseoptions" => {
-                refuse = value.split_whitespace().map(|s| s.to_string()).collect();
-            }
-            _ => {
-                return Err(EngineError::Other(format!(
-                    "unknown daemon parameter: {name}"
-                )));
-            }
-        }
-    }
-
-    if numeric_ids_flag {
-        for m in modules.values_mut() {
-            m.numeric_ids = true;
-        }
-    }
-    if let Some(val) = read_only {
-        for m in modules.values_mut() {
-            m.read_only = val;
-        }
-    }
-    if !refuse.is_empty() {
-        for m in modules.values_mut() {
-            m.refuse_options = refuse.clone();
-        }
-    }
-
-    let addr_family = if matches.get_flag("ipv4") {
-        Some(AddressFamily::V4)
-    } else if matches.get_flag("ipv6") {
-        Some(AddressFamily::V6)
-    } else {
-        None
-    };
-
-    let handler: Arc<daemon::Handler> = Arc::new(|_| Ok(()));
-    let quiet = matches.get_flag("quiet");
-
-    daemon::run_daemon(
-        modules,
-        secrets,
-        password,
-        hosts_allow,
-        hosts_deny,
-        log_file,
-        log_format,
-        motd,
-        pid_file,
-        lock_file,
-        state_dir,
-        timeout,
-        bwlimit,
-        max_conn,
-        refuse,
-        list,
-        port,
-        address,
-        addr_family,
-        65534,
-        65534,
-        handler,
-        quiet,
-    )
-    .map_err(|e| EngineError::Other(format!("daemon failed to bind to port {port}: {e}")))
-}
-
 fn run_probe(opts: ProbeOpts, quiet: bool) -> Result<()> {
     if let Some(addr) = opts.addr {
         let mut stream = TcpStream::connect(&addr)?;
@@ -2597,8 +1258,10 @@ fn run_probe(opts: ProbeOpts, quiet: bool) -> Result<()> {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::utils::{parse_bool, parse_remote_spec, RemoteSpec};
     use daemon::authenticate;
     use engine::SyncOptions;
+    use std::path::PathBuf;
 
     #[test]
     fn windows_paths_are_local() {
